пофиксил черный экран, в appdelegate перенес фичи
