+++ HomeWProject/general /SceneDelegate.swift
@@ -1,0 +1,26 @@
+//
+//  SceneDelegate.swift
+//  HomeWProject
+//
+//  Created by Шарон on 13.08.2025.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+    var window: UIWindow?
+    
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        // Используйте этот метод для настройки окна сцены
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        
+        window = UIWindow(windowScene: windowScene)
+        
+        // Создаем наш ViewController
+        let viewController = ViewController()
+        window?.rootViewController = viewController
+        window?.makeKeyAndVisible()
+    }
+    
+    // Остальные методы по необходимости
+}
